Added support for .MAPGEO version 7
Version 7 was found on the PBE for patch 9.16, and all shipping .mapgeos at the time were updated to this new version (including out-of-use files, specifically Butcher's Bridge and the Trials SR skin).

All of the files at this time have been verified as working perfectly with the converter under this new version.